tests/wpt/web-platform-tests/html/canvas/offscreen/line-styles/2d.line.join.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.line.join.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.join.invalid</h1>
<p class="desc">Setting lineJoin to invalid values is ignored</p>


<script>
var t = async_test("Setting lineJoin to invalid values is ignored");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.lineJoin = 'bevel'
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = 'invalid';
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = 'ROUND';
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = 'round\0';
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = 'round ';
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = "";
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
ctx.lineJoin = 'bevel';
ctx.lineJoin = 'butt';
_assertSame(ctx.lineJoin, 'bevel', "ctx.lineJoin", "'bevel'");
t.done();

});
</script>
